re-try delete to compensate possible race conditions

+++ redaxo/src/core/lib/util/file.php
@@ -195,7 +195,15 @@ class rex_file
     {
         return rex_timer::measure(__METHOD__, static function () use ($file) {
             if (is_file($file)) {
-                return unlink($file);
+                $tryUnlink = @unlink($file);
+
+                // re-try without error suppression to compensate possible race conditions
+                if (!$tryUnlink) {
+                    clearstatcache($file);
+                    if (is_file($file)) {
+                        return unlink($file);
+                    }
+                }                
             }
             return true;
         });
